requirements.txt
# Common
click
pandas
openpyxl
sqlalchemy
psycopg2-binary

# Testing
pytest
pytest_asyncio
testcontainers

# Streamlit
streamlit

# Yandex GPT
yandex-ai-studio-sdk[langchain]

# Langchain / Langgraph
langgraph
langchain_core
langchain-community
langchain_experimental
langchain-ollama
langchain-deepseek

# Pydantic AI
pydantic-ai-slim[openai,logfire,cli,mcp]
# database-pydantic-ai
# database-pydantic-ai @ git+https://github.com/skolchin/database-pydantic-ai.git

# Semantic layer
# malloy
# boring-semantic-layer[agent] @ git+https://github.com/skolchin/boring-semantic-layer.git

